Reverting to a simpler layout
Initially, the plan was to use a complex view for ingredients. Where
user can add a new ingredient set by pressing + button. But, the design
of the system to add dynamic views was very complex. If I were to do
this with Jetpack Compose, it would have been much much easier. So, for
noww, to meet the submission deadline, I'll have to roll it back and use
a simple edit text view for adding ingredients. Focus will be to make
the base implementation complete. All else can be revisited later.
This commit marks the point from where I can revert back to when
revisiting the old implementation.